compressor: introducing new compression methods.
Not using dirtools anymore, because it didn't provide methods for decompression.
It seems I'll have to start writing all of the compression/decompression functions by myself - with the help of python's tarfile, of course.

Archives will remain tar, but compression mode is now bz2:
tar usually compresses better with bzip2 than with gzip.